Moves saving notes to addNote()
Prior to this commit, the application on browser refresh seemingly did
not read localStorage! On browser refresh, there was never any notes
displayed! I found this post:

https://stackoverflow.com/questions/72222728/why-is-localstorage-getting-cleared-whenever-i-refresh-the-page

It pointed out that two useEffects() can be confusing, and suggested
reading localStorage when doing the initial useState. I didn't do this,
but based on the post I removed the useEffect() that saved the notes,
and moved that code into addNote(). Now there is only one useEffect()
which runs only when App is 'mounted', per:

https://react.dev/learn/synchronizing-with-effects

That react.dev post also has another suggestion for how to load the
data, and that is 'outside' the App(). I'll try that next. For now,
this code works!