Revamps notes retrieval from localStorage
The big change is that index.js retrieves the notes from
localStorage. When index.js instantiates  <App/>, it passes the notes
into the App via a property!

To support this, I remove the useEffect() that retrieves the
localStorage, and I bring back the useEffect that sets the localStorage
whenever the notes array changes.

This feels better to me. Per:

https://react.dev/learn/synchronizing-with-effects#not-an-effect-initializing-the-application

"Some logic should only run once when the application starts. You can
put can put it outside your components... This guarantees that such
logic only runs once after the browser loads the page."